Add breakpoints and stack trace events, model

+++ src/main/kotlin/allegrex/agent/ppsspp/bridge/model/PpssppMemoryRange.kt
@@ -4,6 +4,6 @@ data class PpssppMemoryRange(
   val type: String,
   val subtype: String,
   val name: String,
-  val address: Int,
-  val size: Int
+  val address: Long,
+  val size: Long
 ) : PpssppModelKey
